Restructure the server code using blueoprint
