+++ lagrange_server.py
@@ -1,0 +1,4 @@
+from backend import app
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0", port=5002)
