+++ requirements.txt
@@ -1,16 +1,25 @@
-flask
+flask~=2.2.2
 flask-sqlalchemy
 flask_jwt_extended
 web3==5.31.1
-MarkupSafe==2.0.1
+MarkupSafe==2.1.1
 mock
 gevent
 gunicorn
 flask_sslify
 PyMySQL
-flask_script
 Flask-Cors==3.0.10
-werkzeug
-cryptography
-python-mcs-sdk==0.1.5
+werkzeug==2.2.2
+cryptography==38.0.4
 python-dotenv~=0.21.0
+jupyter~=1.0.0
+pip==22.3.1
+attrs==22.2.0
+tornado~=6.2
+Jinja2~=3.1.2
+setuptools==65.6.3
+packaging==22.0
+pyparsing~=3.0.9
+coverage~=7.0.1
+waitress~=2.1.2
+python-mcs-sdk==0.1.7
