--- templates/index.html
@@ -1,12 +1,0 @@
-<html>
-<body>
-Hey there.
-<br><div id="msg">Please Enable Javascript</div>
-<br><button onclick='login()'>LOGIN</button>
-
-<a href="secret">Hey, wanna know a secret?</a>
-
-<script src="static/ethers-v4.min.js"></script>
-<script src="static/app.js"></script>
-</body>
-</html>
