--- templates/submit.html
@@ -1,1 +1,0 @@
-submit please
