+++ web/.env.test
@@ -1,10 +1,10 @@
 NODE_ENV = test
 
 # API Address
-VUE_APP_BASEAPI = 'http://18.221.71.211:5000/'
+VUE_APP_BASEAPI = 'http://127.0.0.1:5002/'
 
 # Signature Domain Name
-VUE_APP_DOMAINNAME = 'lagrangedao.org'
+VUE_APP_DOMAINNAME = '127.0.0.1'
 
 # Fevm RPC
 VUE_APP_FEVMRPC = 'https://wallaby.node.glif.io/rpc/v0'
